fix: 위도 경도 자료형 수정

+++ src/main/java/kr/kro/photoliner/domain/photo/dto/request/CreatePhotosRequest.java
@@ -32,12 +32,12 @@ public record CreatePhotosRequest(
             @Nullable
             @Min(0)
             @Max(90)
-            Integer latitude,
+            Double latitude,
 
             @Nullable
             @Min(0)
             @Max(180)
-            Integer longitude
+            Double longitude
     ) {
 
         public Coordinate convertToGeo() {
